refactor: gateway를 app, chat, room으로 분리
- app: 연결/해제/로그아웃
- chat: 채팅 관련
- room: 나머지 방 관련
