+++ be/src/app.module.ts
@@ -4,8 +4,10 @@ import { ConfigModule } from '@nestjs/config';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { AppController } from './app.controller';
 import { AppService } from './app.service';
+import { AppGateway } from './app.gateway';
 import databaseConfig from './providers/database/database.config';
 import { ChatModule } from './modules/chat/chat.module';
+import { RoomModule } from './modules/room/room.module';
 import { AuthModule } from './modules/auth/auth.module';
 import { RedisModule } from './providers/redis/redis.module';
 import { TransformInterceptor } from './common/interceptors/transform.interceptor';
@@ -17,11 +19,13 @@ import { HttpExceptionFilter } from './common/filters/http-exception.filter';
     TypeOrmModule.forRoot({ ...databaseConfig, entities: [], migrations: [] }),
     RedisModule,
     ChatModule,
+    RoomModule,
     AuthModule,
   ],
   controllers: [AppController],
   providers: [
     AppService,
+    AppGateway,
     {
       provide: APP_INTERCEPTOR,
       useClass: TransformInterceptor,
