+++ be/src/modules/chat/chat.module.ts
@@ -1,11 +1,11 @@
 import { Module } from '@nestjs/common';
 import { ChatService } from './chat.service';
 import { ChatGateway } from './chat.gateway';
-import { RoomsModule } from '@src/modules/room/room.module';
+import { RoomModule } from '@src/modules/room/room.module';
 import { AuthModule } from '@src/modules/auth/auth.module';
 
 @Module({
-  imports: [RoomsModule, AuthModule],
+  imports: [RoomModule, AuthModule],
   providers: [ChatService, ChatGateway],
   exports: [ChatService],
 })
