+++ be/src/modules/room/room.module.ts
@@ -2,11 +2,12 @@ import { Module } from '@nestjs/common';
 import { AuthModule } from '../auth/auth.module';
 import { RoomController } from './room.controller';
 import { RoomService } from './room.service';
+import { RoomGateway } from './room.gateway';
 
 @Module({
   imports: [AuthModule],
   controllers: [RoomController],
-  providers: [RoomService],
+  providers: [RoomService, RoomGateway],
   exports: [RoomService],
 })
-export class RoomsModule {}
+export class RoomModule {}
